Changed markers in the CustomMapDisplay Improved MapMarker performance Moved image utils to the same file
