--- src/components/battle-map/custom-map-dialog/TargetIcon.ts
@@ -1,10 +1,0 @@
-import { Icon } from 'leaflet';
-
-import _TargetIcon from '../../../assets/target.png';
-
-const TargetIcon = new Icon({
-  iconUrl: _TargetIcon,
-  iconSize: [35, 35],
-});
-
-export default TargetIcon;
