--- src/utils/canvas.ts
@@ -1,52 +1,0 @@
-// https://codesandbox.io/s/react-image-crop-demo-with-react-hooks-y831o
-import { PixelCrop } from 'react-image-crop';
-
-// eslint-disable-next-line import/prefer-default-export
-export function canvasPreview(
-  image: HTMLImageElement,
-  canvas: HTMLCanvasElement,
-  crop: PixelCrop,
-) {
-  const ctx = canvas.getContext('2d');
-
-  if (!ctx) {
-    throw new Error('No 2d context');
-  }
-
-  const scaleX = image.naturalWidth / image.width;
-  const scaleY = image.naturalHeight / image.height;
-  // devicePixelRatio slightly increases sharpness on retina devices
-  // at the expense of slightly slower render times and needing to
-  // size the image back down if you want to download/upload and be
-  // true to the images natural size.
-  const pixelRatio = window.devicePixelRatio;
-  // const pixelRatio = 1
-
-  /* eslint-disable no-param-reassign */
-  canvas.width = Math.floor(crop.width * scaleX * pixelRatio);
-  canvas.height = Math.floor(crop.height * scaleY * pixelRatio);
-  /* eslint-enable no-param-reassign */
-
-  ctx.scale(pixelRatio, pixelRatio);
-  ctx.imageSmoothingQuality = 'high';
-
-  const cropX = crop.x * scaleX;
-  const cropY = crop.y * scaleY;
-
-  ctx.save();
-  // Move the crop origin to the canvas origin (0,0)
-  ctx.translate(-cropX, -cropY);
-  // Draw it
-  ctx.drawImage(
-    image,
-    0,
-    0,
-    image.naturalWidth,
-    image.naturalHeight,
-    0,
-    0,
-    image.naturalWidth,
-    image.naturalHeight,
-  );
-  ctx.restore();
-}
